# docs/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Antaride",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/v1/driver/orders/{uuid}/status": {
            "patch": {
                "operationId": "v1.driver.orders.transition",
                "description": "Satu endpoint untuk seluruh transisi \u2014 menuju penjemputan, tiba, mulai \u2014\nbukan satu endpoint per transisi. Alasannya: aturan transisi yang sah ada\ndi state machine, dan endpoint terpisah per transisi berarti daftar aturan\nyang sama tersebar di beberapa route.",
                "summary": "Ubah status order yang sedang dipegang",
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransitionOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "anyOf": [
                                                {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/v1/driver/orders/{uuid}/complete": {
            "post": {
                "operationId": "v1.driver.orders.complete",
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompleteOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "anyOf": [
                                                {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/v1/auth/otp/request": {
            "post": {
                "operationId": "v1.auth.otp.request",
                "description": "Response-nya SELALU sama bentuknya, terdaftar maupun belum.",
                "summary": "Minta kode OTP",
                "tags": [
                    "Otp"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestOtpRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "anyOf": [
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "phone_masked": {
                                                            "type": "string",
                                                            "description": "Nomor dikembalikan dalam bentuk tersamarkan, bukan penuh.\nAplikasi sudah tahu nomor yang dia kirim; yang ditampilkan di\nlayar \"masukkan kode\" adalah konfirmasi, dan nomor penuh di sana\nhanya menambah data yang bisa dibaca orang di sekitar."
                                                        },
                                                        "purpose": {
                                                            "type": "string"
                                                        },
                                                        "expires_at": {
                                                            "type": "string"
                                                        },
                                                        "expires_in_seconds": {
                                                            "type": [
                                                                "object",
                                                                "null"
                                                            ]
                                                        },
                                                        "resend_after_seconds": {
                                                            "type": "integer"
                                                        },
                                                        "debug_code": {
                                                            "type": [
                                                                "string",
                                                                "null"
                                                            ]
                                                        }
                                                    },
                                                    "required": [
                                                        "phone_masked",
                                                        "purpose",
                                                        "expires_at",
                                                        "expires_in_seconds",
                                                        "resend_after_seconds",
                                                        "debug_code"
                                                    ]
                                                },
                                                {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompleteOrderRequest": {
                "type": "object",
                "description": "Driver menyelesaikan order.",
                "properties": {
                    "lat": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                    },
                    "lng": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                    },
                    "actual_polyline": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Jejak GPS perjalanan, dikirim sebagai polyline terenkode. Opsional karena bisa hilang: aplikasi yang dimatikan paksa di\ntengah perjalanan kehilangan bufernya, dan menolak penyelesaian\norder karena itu berarti perjalanan yang sudah selesai tidak bisa\nditutup \u2014 dan driver terjebak, karena partial unique index\nmelarangnya menerima order berikutnya.",
                        "maxLength": 65535
                    },
                    "actual_distance_m": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "Jarak yang dilaporkan aplikasi. Diterima, tapi hanya dipakai kalau polyline tidak ada. Angka ini\nyang paling menarik untuk dipalsukan \u2014 jarak lebih besar berarti\norder masuk antrean review dan ongkosnya bisa dinaikkan \u2014 jadi\nyang lebih dipercaya adalah jejak titiknya. Lihat CompleteOrder.",
                        "minimum": 0,
                        "maximum": 2000000
                    }
                },
                "required": [
                    "lat",
                    "lng"
                ],
                "title": "CompleteOrderRequest"
            },
...
            "TransitionOrderRequest": {
                "type": "object",
                "description": "Ubah status order oleh driver.\n\n============================================================================\n KOORDINAT WAJIB, DAN ITU PILIHAN YANG DISENGAJA\n============================================================================\n Setiap transisi status oleh driver harus disertai posisinya. Itu satu-satunya\n cara menjawab pertanyaan yang paling sering muncul dalam sengketa:\n\n     \"Apakah driver benar-benar ada di titik jemput saat menekan 'tiba'?\"\n\n Tanpa koordinat, jawabannya tidak ada di mana pun, dan yang tersisa hanya\n keterangan dua pihak yang bertentangan.\n\n Posisi dari GPS bisa dipalsukan dengan aplikasi mock location, dan koordinat\n ini TIDAK diperlakukan sebagai bukti mutlak. Tapi bedanya besar: memalsukan\n posisi menuntut usaha yang terlihat di data (lompatan posisi, akurasi\n sempurna yang tidak wajar), sementara tanpa data sama sekali tidak ada yang\n bisa diperiksa.\n============================================================================",
                "properties": {
                    "status": {
                        "type": "string",
                        "description": "Status tujuan dibatasi ke transisi yang MASUK AKAL dari aplikasi\ndriver. `completed` dan `cancelled` sengaja TIDAK ada di sini: keduanya\npunya endpoint sendiri karena membawa efek samping yang jauh lebih\nbesar \u2014 settlement uang dan pelepasan dana tertahan. Menyatukannya\nke endpoint transisi generik berarti satu request yang salah\nstatus bisa memicu pembagian uang.",
                        "enum": [
                            "driver_arriving",
                            "driver_arrived"
                        ]
                    },
                    "lat": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                    },
                    "lng": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                    },
                    "accuracy_m": {
                        "type": [
                            "number",
                            "null"
                        ],
                        "description": "Akurasi dikirim supaya panel admin bisa menilai seberapa layak\nposisinya dipercaya saat ada sengketa.",
                        "minimum": 0,
                        "maximum": 10000
                    }
                },
                "required": [
                    "status",
                    "lat",
                    "lng"
                ],
                "title": "TransitionOrderRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}